ModuleDef and schema changes to support procedures
This commit extends various schema and schema-adjacent structures to describe procedures,
a new kind of database function which are allowed to perform side effects.

This includes extending `RawModuleDefV9` with a way to register `RawProcedureDefV9`s
in the `misc_exports`, preserving compatibility with modules that predate procedures.

The module validation path is reorganized somewhat
to validate various properties related to procedures
while preserving code clarity and maintainability.

Additionally, the `ArgsTuple` machinery for ser/de-ing reducer arguments
using the argument type as a seed is extended to also support procedure arguments.

All of this is currently unused.