Rust module bindings and macros for defining procedures
This commit adds a macro attribute `#[procedure]` which applies to functions,
and various in-WASM machinery for defining, calling and running procedures.

A simple example of a procedure, included in `modules/module-test`:

```rust
fn sleep_one_second(ctx: &mut ProcedureContext) {
    let prev_time = ctx.timestamp;
    let target = prev_time + Duration::from_secs(1);
    ctx.sleep_until(target);
    let new_time = ctx.timestamp;
    let actual_delta = new_time.duration_since(prev_time).unwrap();
    log::info!("Slept from {prev_time} to {new_time}, a total of {actual_delta:?}");
}
```

We intend eventually to make procedures be `async` functions
(with the trivial `now_or_never` executor from `future-util`),
but I found that making the types work for this was giving me a lot of trouble,
and decided to put it off in the interest of unblocking more parallelizable work.

Host-side infrastructure for executing procedures is not included in this commit.
I have a prototype working, but cleaning it up for review and merge will come a bit later.