[TS] Fix ctx.timestamp serialization (#3461)
# Description of Changes

Resolves #3433. The issue was caused by the fact we were calling
`serialize_to_js` with a full `Timestamp` struct, not just an `i64`. In
the future I'd like to move away from the use of `serialize_to_js` to
avoid problems like this.

# API and ABI breaking changes

Technically a breaking change, but fixes a bug.

# Expected complexity level and risk

1

# Testing

<!-- Describe any testing you've done, and any testing you'd like your
reviewers to do,
so that you're confident that all the changes work as expected! -->

- [x] Checked to ensure `ctx.timestamp` is correct.
- [ ] <!-- maybe a test you want a reviewer to do, so they can check it
off when they're satisfied. -->