WASM host execution for procedures
This commit builds support for executing procedures in WASM modules.
This includes an HTTP endpoint, `/v1/database/:name_or_address/procedure/:name POST`,
as well as an extension to the WS protocol.
These new APIs are not wired up to the CLI or SDKs,
but I have manually tested the HTTP endpoint via `curl`.
The new WS extensions are completely untested.

Several TODOs are scattered throughout the new code,
most notably for sensibly tracking procedure execution time in the metrics.

I also expect that we will want to remove the `procedure_sleep_until` syscall
and the `ProcedureContext::sleep_until` method prior to release.