Fix to Unreal tests to stop requesting infinite resources (#3504)
# Description of Changes

For core dumps Unreal tries to use setrlimit() for infinity when setting
resource limits, it seems this becomes an issue intermittently when
running tests in parallel. There is an option when running Unreal to
disable this and we can rely on error longs only which makes sense for
CI imo.

# API and ABI breaking changes

None

# Expected complexity level and risk

1 - small flag change

# Testing

- [x] Re-ran all tests locally