Fix Nix build by not using Git in Cargo build scripts
When building under Nix, Git metadata is not available within the sandbox,
as we use `lib.cleanSource` on our source directory.
This is important because it avoids spurious rebuilds and/or determinism hazards.

The build was broken due to our new `spacetime init` template system
accessing Git metadata in the CLI's build.rs
to filter out non-git-tracked files from the templates.
The Flake sandbox does this automatically (even without `lib.cleanSource`!),
so when building under Nix it's unnecessary to do twice.
(I remain unconvinced that it's necessary to do in non-Nix builds either,
as CI builds should have a clean checkout and local dev builds don't need clean templates,
but the behavior was already in master and I didn't feel comfortable removing it.)

As an enhancement, I've also found a Nix-ey way to embed our Git commit hash in builds.
Previously, builds under Nix had the empty string instead of a commit hash,
because we included the `git` CLI tool but scrubbed the necessary metadata.
Now, we inject an environment variable from the Nix flake,
and don't make the `git` CLI tool available at all.
This has the convenient upside of allowing Nix builds to reference `dirtyRev`
in builds with a dirty worktree, which should reduce confusion.