Use production API key (#3560)
# Description of Changes

Inkeep's API key for the search bar is the staging one instead of
production, therefore requests made are failing due to an incorrect
referrer.
I've fixed this on `docs/release` quickly after we migrated the new docs
but haven't made a PR to add it to master, due to this the change was
lost on `docs/release` in the last docs release.
I just cherry-picked the commit that I had on my local branch

# API and ABI breaking changes

None.

# Expected complexity level and risk

1

# Testing

Cherry-pick of a commit that was already deployed