Add proper V8 sys module versioning (#3527)
# Description of Changes

The previous `syscall.rs` is now in `syscall/v1.rs`, and there's proper
parsing of the `sys@X.Y` version number done now. This lets us more
easily extend the ABI in the future.

I also changed how slots are stored in the Context embedder data, to
make it easier to add more in the future and more performant to fetch
them.

# Expected complexity level and risk

1

# Testing

n/a, reorganization

---------

Signed-off-by: Noa <coolreader18@gmail.com>
Co-authored-by: Mazdak Farrokhzad <twingoow@gmail.com>
Co-authored-by: Zeke Foppa <bfops@users.noreply.github.com>