Refactor module instance code (#3605)
# Description of Changes

The main thing this does is implement `WasmInstance` for `V8Instance`,
so we can now change all the `vm_call*: impl FnOnce()` args on
`InstanceCommon` to instead just take `&mut impl WasmInstance`.
(Obviously our javascript instance is not a wasm instance, but
`wasmtime::module_host_actor` is desperately in need of a rebrand in
general, so I just left `WasmInstance`'s as is.)
 
This vastly simplifies the callsites, and also allows @joshua-spacetime
to call views from inside of `call_reducer_with_tx`, which was the main
motivating factor behind this.

If the last commit is a bit much, I'm happy to split it off - I just
don't get explicitly asked to refactor the rough edges of the codebase
very often :3

# Expected complexity level and risk

2 - it's a refactor, but overall pretty straightforward, and Rust lets
you do fearless refactoring etc etc

# Testing

n/a - internal code change

---------

Signed-off-by: Noa <coolreader18@gmail.com>
Co-authored-by: joshua-spacetime <josh@clockworklabs.io>