Update 03-part-2.md
Fixed client module bindings autogen location in a CLI call. 
The old command generates the bindings outside of the Unity project's actual Assets folder and they can't be seen by the GameManager script. It's also inconsistent with the project structure presented at the beginning of part 2.
The new command is consistent with part 3 of the tutorial, where the bindings are being regenerated.

Signed-off-by: T-Podgorski <147391857+T-Podgorski@users.noreply.github.com>